+++ changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Unreleased
+Features being worked on.
+
+### Added
++ Quick-edit box for selected objects
++ Arrow-key + modifiers for controlling quick-edit box inputs
+
+### Fixed
++ Allow selection of objects when not using the selection tool, by switching to the selection tool
++ List should keep scroll position if re-rendered
+
 ## [0.1.1] - 2020-12-03
 Fixes for some small annoyances.